9043/refactor: extract some method from UBSClientService to OrderCheckoutService
