+++ service-api/src/main/java/greencity/service/ubs/point/PointService.java
@@ -1,0 +1,4 @@
+package greencity.service.ubs.point;
+
+public interface PointService {
+}
