+++ service-api/src/main/java/greencity/service/ubs/tariff/TariffService.java
@@ -1,0 +1,4 @@
+package greencity.service.ubs.tariff;
+
+public interface TariffService {
+}
