+++ service-api/src/main/java/greencity/service/ubs/user/CourierService.java
@@ -1,0 +1,4 @@
+package greencity.service.ubs.user;
+
+public interface CourierService {
+}
