+++ service-api/src/main/java/greencity/service/ubs/user/UserService.java
@@ -1,5 +1,4 @@
 package greencity.service.ubs.user;
 
-public interface UbsUserService {
-
+public interface UserService {
 }
