+++ service/src/main/java/greencity/service/ubs/UBSClientServiceImpl.java
@@ -10,15 +10,11 @@ import static greencity.constant.ErrorMessage.COURIER_IS_NOT_FOUND_BY_ID;
 import static greencity.constant.ErrorMessage.EMPLOYEE_DOESNT_EXIST;
 import static greencity.constant.ErrorMessage.EVENTS_NOT_FOUND_EXCEPTION;
 import static greencity.constant.ErrorMessage.LOCATION_DOESNT_FOUND_BY_ID;
-import static greencity.constant.ErrorMessage.LOCATION_IS_DEACTIVATED_FOR_TARIFF;
 import static greencity.constant.ErrorMessage.ORDER_WITH_CURRENT_ID_DOES_NOT_EXIST;
 import static greencity.constant.ErrorMessage.RECIPIENT_WITH_CURRENT_ID_DOES_NOT_EXIST;
 import static greencity.constant.ErrorMessage.TARIFF_FOR_COURIER_AND_LOCATION_NOT_EXIST;
-import static greencity.constant.ErrorMessage.TARIFF_FOR_LOCATION_NOT_EXIST;
 import static greencity.constant.ErrorMessage.TARIFF_FOR_ORDER_NOT_EXIST;
-import static greencity.constant.ErrorMessage.TARIFF_NOT_FOUND;
 import static greencity.constant.ErrorMessage.TARIFF_NOT_FOUND_BY_LOCATION_ID;
-import static greencity.constant.ErrorMessage.TARIFF_OR_LOCATION_IS_DEACTIVATED;
 import static greencity.constant.ErrorMessage.USER_WITH_CURRENT_ID_DOES_NOT_EXIST;
 import static greencity.constant.ErrorMessage.USER_WITH_CURRENT_UUID_ALREADY_EXISTS_IN_UBS;
 import static greencity.constant.ErrorMessage.USER_WITH_CURRENT_UUID_DOES_NOT_EXIST;
@@ -27,7 +23,6 @@ import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
 import greencity.client.UserRemoteClient;
 import greencity.constant.AppConstant;
-import greencity.constant.ErrorMessage;
 import greencity.constant.OrderHistory;
 import greencity.dto.AllActiveLocationsDto;
 import greencity.dto.LocationWithTariffInfoDto;
@@ -40,7 +35,6 @@ import greencity.dto.TariffsForLocationDto;
 import greencity.dto.address.AddressDto;
 import greencity.dto.address.AddressInfoDto;
 import greencity.dto.bag.BagForUserDto;
-import greencity.dto.bag.BagTranslationDto;
 import greencity.dto.certificate.CertificateDto;
 import greencity.dto.courier.CourierDto;
 import greencity.dto.customer.UbsCustomersDto;
@@ -56,15 +50,12 @@ import greencity.dto.pageble.PageableDto;
 import greencity.dto.position.PositionAuthoritiesDto;
 import greencity.dto.user.AllPointsUserDto;
 import greencity.dto.user.DeactivateUserRequestDto;
-import greencity.dto.user.PersonalDataDto;
 import greencity.dto.user.PointsForUbsUserDto;
 import greencity.dto.user.UserInfoDto;
 import greencity.dto.user.UserPointDto;
-import greencity.dto.user.UserPointsAndAllBagsDto;
 import greencity.dto.user.UserProfileCreateDto;
 import greencity.dto.user.UserProfileDto;
 import greencity.dto.user.UserProfileUpdateDto;
-import greencity.entity.order.Bag;
 import greencity.entity.order.Certificate;
 import greencity.entity.order.ChangeOfPoints;
 import greencity.entity.order.Event;
@@ -82,9 +73,7 @@ import greencity.entity.user.ubs.OrderAddress;
 import greencity.entity.user.ubs.UBSuser;
 import greencity.enums.BotType;
 import greencity.enums.CertificateStatus;
-import greencity.enums.LocationStatus;
 import greencity.enums.OrderStatus;
-import greencity.enums.TariffStatus;
 import greencity.exceptions.BadRequestException;
 import greencity.exceptions.NotFoundException;
 import greencity.exceptions.http.AccessDeniedException;
@@ -92,18 +81,15 @@ import greencity.exceptions.user.UBSuserNotFoundException;
 import greencity.exceptions.user.UserNotFoundException;
 import greencity.mapping.location.LocationToLocationsDtoMapper;
 import greencity.repository.AddressRepository;
-import greencity.repository.BagRepository;
 import greencity.repository.CertificateRepository;
 import greencity.repository.CourierRepository;
 import greencity.repository.EmployeeRepository;
 import greencity.repository.EventRepository;
 import greencity.repository.LocationRepository;
-import greencity.repository.OrderBagRepository;
 import greencity.repository.OrderPaymentStatusTranslationRepository;
 import greencity.repository.OrderRepository;
 import greencity.repository.OrderStatusTranslationRepository;
 import greencity.repository.OrdersForUserRepository;
-import greencity.repository.TariffLocationRepository;
 import greencity.repository.TariffsInfoRepository;
 import greencity.repository.TelegramChatRepository;
 import greencity.repository.UBSUserRepository;
@@ -115,12 +101,10 @@ import greencity.service.ubs.calculator.PaymentCalculatorService;
 import greencity.util.Bot;
 import greencity.util.MoneyConverterUtil;
 import jakarta.transaction.Transactional;
-import java.math.BigDecimal;
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Base64;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.EnumSet;
 import java.util.List;
@@ -150,7 +134,6 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class UBSClientServiceImpl implements UBSClientService {
     private final UserRepository userRepository;
-    private final BagRepository bagRepository;
     private final UBSUserRepository ubsUserRepository;
     private final ModelMapper modelMapper;
     private final CertificateRepository certificateRepository;
@@ -164,11 +147,9 @@ public class UBSClientServiceImpl implements UBSClientService {
     private final OrderStatusTranslationRepository orderStatusTranslationRepository;
     private final OrderPaymentStatusTranslationRepository orderPaymentStatusTranslationRepository;
     private final EventService eventService;
-    private final TariffLocationRepository tariffLocationRepository;
     private final LocationRepository locationRepository;
     private final TariffsInfoRepository tariffsInfoRepository;
     private final TelegramChatRepository telegramBotRepository;
-    private final OrderBagRepository orderBagRepository;
     private final LocationToLocationsDtoMapper locationToLocationsDtoMapper;
     private final AddressService addressService;
     private final PaymentCalculatorService paymentCalculatorService;
@@ -211,131 +192,6 @@ public class UBSClientServiceImpl implements UBSClientService {
         String decodedString = new String(decodedBytes, StandardCharsets.UTF_8);
         JSONObject jsonObject = new JSONObject(decodedString);
         return jsonObject.getString("status");
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public UserPointsAndAllBagsDto getFirstPageDataByTariffAndLocationId(Long tariffId, Long locationId) {
-        TariffsInfo tariffsInfo = tariffsInfoRepository.findById(tariffId)
-            .orElseThrow(() -> new NotFoundException(TARIFF_NOT_FOUND + tariffId));
-
-        Location location = locationRepository.findById(locationId)
-            .orElseThrow(() -> new NotFoundException(LOCATION_DOESNT_FOUND_BY_ID + locationId));
-
-        checkIfTariffIsAvailableForCurrentLocation(tariffsInfo, location);
-
-        return getUserPointsAndAllBagsDtoByTariffIdAndUserPoints(tariffsInfo.getId(), 0);
-    }
-
-    @Override
-    public UserPointsAndAllBagsDto getFirstPageDataByOrderId(String uuid, Long orderId) {
-        User user = userRepository.findUserByUuid(uuid).orElseThrow(
-            () -> new NotFoundException(USER_WITH_CURRENT_UUID_DOES_NOT_EXIST));
-        Order order = orderRepository.findById(orderId).orElseThrow(
-            () -> new NotFoundException(ORDER_WITH_CURRENT_ID_DOES_NOT_EXIST + orderId));
-
-        checkIsOrderOfCurrentUser(user, order);
-
-        TariffsInfo tariffsInfo = order.getTariffsInfo();
-
-        Location location = getLocationByOrderIdThroughLazyInitialization(order);
-
-        checkIfTariffIsAvailableForCurrentLocation(tariffsInfo, location);
-
-        return getUserPointsAndAllBagsDtoByTariffIdAndOrderIdAndUserPoints(tariffsInfo.getId(), user.getCurrentPoints(),
-            orderId);
-    }
-
-    private void checkIsOrderOfCurrentUser(User user, Order order) {
-        if (!order.getUser().getId().equals(user.getId())) {
-            throw new AccessDeniedException(ErrorMessage.ORDER_DOES_NOT_BELONG_TO_USER);
-        }
-    }
-
-    private void checkIfTariffIsAvailableForCurrentLocation(TariffsInfo tariffsInfo, Location location) {
-        if (tariffsInfo.getTariffStatus() == TariffStatus.DEACTIVATED
-            || location.getLocationStatus() == LocationStatus.DEACTIVATED) {
-            throw new BadRequestException(TARIFF_OR_LOCATION_IS_DEACTIVATED);
-        } else {
-            boolean isAvailable = isTariffAvailableForCurrentLocation(tariffsInfo, location);
-            if (!isAvailable) {
-                throw new BadRequestException(LOCATION_IS_DEACTIVATED_FOR_TARIFF + tariffsInfo.getId());
-            }
-        }
-    }
-
-    private boolean isTariffAvailableForCurrentLocation(TariffsInfo tariffsInfo, Location location) {
-        return tariffLocationRepository
-            .findTariffLocationByTariffsInfoAndLocation(tariffsInfo, location)
-            .orElseThrow(() -> new NotFoundException(TARIFF_FOR_LOCATION_NOT_EXIST + location.getId()))
-            .getLocationStatus() != LocationStatus.DEACTIVATED;
-    }
-
-    private UserPointsAndAllBagsDto getUserPointsAndAllBagsDtoByTariffIdAndOrderIdAndUserPoints(Long tariffId,
-        Integer userPoints, Long orderId) {
-        List<BagTranslationDto> bagTranslationDtoList =
-            bagRepository.findAllActiveBagsByTariffsInfoId(tariffId).stream()
-                .map(bag -> buildBagTranslationDto(orderId, bag))
-                .toList();
-        return new UserPointsAndAllBagsDto(bagTranslationDtoList, userPoints);
-    }
-
-    private UserPointsAndAllBagsDto getUserPointsAndAllBagsDtoByTariffIdAndUserPoints(Long tariffId,
-        Integer userPoints) {
-        List<BagTranslationDto> bagTranslationDtoList =
-            bagRepository.findAllActiveBagsByTariffsInfoId(tariffId).stream()
-                .map(bag -> modelMapper.map(bag, BagTranslationDto.class))
-                .sorted(Comparator.comparing(BagTranslationDto::getCapacity).reversed())
-                .toList();
-        return new UserPointsAndAllBagsDto(bagTranslationDtoList, userPoints);
-    }
-
-    private BagTranslationDto buildBagTranslationDto(Long orderId, Bag source) {
-        return BagTranslationDto.builder()
-            .id(source.getId())
-            .capacity(source.getCapacity())
-            .price(BigDecimal.valueOf(source.getFullPrice())
-                .movePointLeft(AppConstant.TWO_DECIMALS_AFTER_POINT_IN_CURRENCY).doubleValue())
-            .nameUk(source.getNameUk())
-            .nameEn(source.getNameEn())
-            .limitedIncluded(source.getLimitIncluded())
-            .quantity(getQuantityOfBagsByBagIdAndOrderId(orderId, source.getId()))
-            .build();
-    }
-
-    private Integer getQuantityOfBagsByBagIdAndOrderId(Long orderId, Integer bagId) {
-        return orderBagRepository.getAmountOfOrderBagsByOrderIdAndBagId(orderId, bagId)
-            .orElse(0);
-    }
-
-    private Location getLocationByOrderIdThroughLazyInitialization(Order order) {
-        return order
-            .getUbsUser()
-            .getOrderAddress()
-            .getLocation();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    @Transactional
-    public PersonalDataDto getSecondPageData(String uuid) {
-        User currentUser = userRepository.findByUuid(uuid);
-        List<UBSuser> ubsUser = ubsUserRepository.findUBSuserByUser(currentUser);
-
-        if (ubsUser.isEmpty()) {
-            ubsUser = Collections.singletonList(UBSuser.builder().id(null).build());
-        }
-        PersonalDataDto dto = modelMapper.map(currentUser, PersonalDataDto.class);
-        dto.setUbsUserId(ubsUser.getFirst().getId());
-        if (currentUser.getAlternateEmail() != null
-            && !currentUser.getAlternateEmail().isEmpty()) {
-            dto.setEmail(currentUser.getAlternateEmail());
-        }
-        return dto;
     }
 
     /**
